my_requirements.txt
# pip install -r my_requirements.txt


diffusers>=0.14.0
accelerate
opencv-contrib-python
transformers
xformers
streamlit
numpy
pillow
controlnet_aux
clip_interrogator
clr
pythonnet
auto_py_to_exe


# torch==2.0.1+cu118
# this is special, need to install manually
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118